Refactor project structure: update package.json, remove reset script, and delete tsconfig.json
- Updated package.json to change the main entry point and streamline scripts.
- Removed the reset-project.js script as it is no longer needed.
- Deleted tsconfig.json to simplify the project setup.
