--- Frontend/STS/app/(tabs)/account.tsx
@@ -1,35 +1,0 @@
-import { Image } from 'expo-image';
-import { Platform, StyleSheet, View } from 'react-native';
-
-import { Collapsible } from '@/components/ui/collapsible';
-import { ExternalLink } from '@/components/external-link';
-import ParallaxScrollView from '@/components/parallax-scroll-view';
-import { ThemedText } from '@/components/themed-text';
-import { ThemedView } from '@/components/themed-view';
-import { IconSymbol } from '@/components/ui/icon-symbol';
-import { Fonts } from '@/constants/theme';
-import { Text } from 'react-native';
-
-
-export default function TabTwoScreen() {
-  return (
-    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center', height: 500 }}>
-      <Text style={{ fontSize: 24, textAlign: 'center' }}>
-        Account
-      </Text>
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  headerImage: {
-    color: '#808080',
-    bottom: -90,
-    left: -35,
-    position: 'absolute',
-  },
-  titleContainer: {
-    flexDirection: 'row',
-    gap: 8,
-  },
-});
