+++ Frontend/STS/package.json
@@ -1,48 +1,18 @@
 {
   "name": "sts",
-  "main": "expo-router/entry",
   "version": "1.0.0",
+  "main": "index.js",
   "scripts": {
     "start": "expo start",
-    "reset-project": "node ./scripts/reset-project.js",
-    "android": "expo run:android",
-    "ios": "expo run:ios",
-    "web": "expo start --web",
-    "lint": "expo lint"
+    "android": "expo start --android",
+    "ios": "expo start --ios",
+    "web": "expo start --web"
   },
   "dependencies": {
-    "@expo/vector-icons": "^15.0.2",
-    "@expo/metro-runtime": "~6.1.2",
-    "@react-navigation/bottom-tabs": "^7.4.0",
-    "@react-navigation/elements": "^2.6.3",
-    "@react-navigation/native": "^7.1.8",
     "expo": "~54.0.10",
-    "expo-constants": "~18.0.9",
-    "expo-font": "~14.0.8",
-    "expo-haptics": "~15.0.7",
-    "expo-image": "~3.0.8",
-    "expo-linking": "~8.0.8",
-    "expo-router": "~6.0.8",
-    "expo-splash-screen": "~31.0.10",
     "expo-status-bar": "~3.0.8",
-    "expo-symbols": "~1.0.7",
-    "expo-system-ui": "~6.0.7",
-    "expo-web-browser": "~15.0.7",
     "react": "19.1.0",
-    "react-dom": "19.1.0",
-    "react-native": "0.81.4",
-    "react-native-gesture-handler": "~2.28.0",
-    "react-native-worklets": "0.5.1",
-    "react-native-reanimated": "~4.1.1",
-    "react-native-safe-area-context": "~5.6.0",
-    "react-native-screens": "~4.16.0",
-    "react-native-web": "~0.21.0"
-  },
-  "devDependencies": {
-    "@types/react": "~19.1.0",
-    "typescript": "~5.9.2",
-    "eslint": "^9.25.0",
-    "eslint-config-expo": "~10.0.0"
+    "react-native": "0.81.4"
   },
   "private": true
 }
